% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_morph.R
\name{adjust_morph}
\alias{adjust_morph}
\title{Remove Allometric Effects}
\usage{
adjust_morph(
  data,
  type = c("multi_pop", "single_pop", "species"),
  char = "SVL"
)
}
\arguments{
\item{data}{A data frame with population or species label in the first column, followed by SVL and other data.}

\item{type}{String. Data type:  "single_pop", "multi_pop" and "species".}

\item{char}{String. A character as size reference (e.g., SVL).}
}
\value{
A data frame containing the adjusted morphological data.
}
\description{
A function to remove the allometric effects of body size in morphological data.
Allometric size correction can be applied to both single and multi-population data, as well as species data.
The \emph{"single_pop"} option considers all data from one species; therefore, an overall mean is utilized to calculate a single slope for the entire dataset.
The \emph{"multi_pop"} option considers all data from one species, while the slope is calculated for each population using an overall mean.
The \emph{"species"} option considers each unique group as a single species; therefore, the slope is calculated for each using their respective mean.
}
\examples{
data <- read.csv(system.file("extdata", "herp.data.csv", package = "MorphoTax"))
data$Sex <- NULL

adj.data <- adjust_morph(data, type = "multi_pop", char = "SVL")

head(adj.data)

}
\references{
Thorpe, R. S. (1975). Quantitative handling of characters useful in snake systematics with particular reference to intraspecific
variation in the Ringed Snake \emph{Natrix natrix} (L.). Biological Journal of the Linnaean Society 7: 27–43. https://doi.org/10.1111/j.1095-8312.1975.tb00732.x
}
