% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_tukey.R
\name{anova_tukey}
\alias{anova_tukey}
\title{ANOVA and Tukey HSD}
\usage{
anova_tukey(data, grp = "Pop", write.tk = FALSE, dir = NULL)
}
\arguments{
\item{data}{A data frame with population or species label in the first column, followed by morpholigcal data.}

\item{grp}{String. Column name for population or species.}

\item{write.tk}{Logical. If TRUE, it writes the results of the Tukey test to csv files.}

\item{dir}{A directory where the Tukey test results will be stored.}
}
\value{
A list containing the results of ANOVA and Tukey Test.
}
\description{
A function to perform ANOVA on all morphological data across populations or species.
Also, if a significant result is detected, it automatically performs Tukey Test.
}
\examples{
data <- read.csv(system.file("extdata", "herp.data.csv", package = "MorphoTax"))
data$Sex <- NULL

aov.res <- anova_tukey(data, grp = "Pop", write.tk = TRUE, dir = NULL)

aov.res$aov_summary
aov.res$tukey_combined

}
