% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_cols.R
\name{arrange_cols}
\alias{arrange_cols}
\title{Arrange Columns}
\usage{
arrange_cols(data, col.names)
}
\arguments{
\item{data}{A data frame with population or species label in the first column, followed by morpholigcal data.}

\item{col.names}{A vector containing the names of the columns to reorder. If column names are not included, those will be dropped.}
}
\value{
A data frame containing the reordered columns.
}
\description{
A function to reorder columns in morphology data, except for the first one i.e., population or species column.
}
\examples{
data <- read.csv(system.file("extdata", "herp.data.csv", package = "MorphoTax"))
data$Sex <- NULL
cols <- c("SVL", "TD", "OD")

data <- arrange_cols(data, col.names = cols)

head(data)

}
