% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_comp.R
\name{assign_comp}
\alias{assign_comp}
\title{Assign Comparisons}
\usage{
assign_comp(data, comp = "comparisons", rules)
}
\arguments{
\item{data}{A data frame with p-values of comparisons.}

\item{comp}{String specifying the column name of comaparisons}

\item{rules}{A list containing the group assignment rules.}
}
\value{
A dataframe with group assignment column.
}
\description{
A function for assigning groups (e.g., species) to each comparison in a dataframe of p-values,
allowing the comparisons to be sorted by groups using the function 'plot_pval_bubble.'
}
\examples{
data <- read.csv(system.file("extdata", "cleaned.pvals.csv", package = "MorphoTax"))
sp.grp <- list(
           marmorata = c("Luzon"),
           philippina = c("Mindanao"),
           mallarii = c("Palawan"))

data <- assign_comp(data, comp = "comparisons", rules = sp.grp)

}
