% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_pvals.R
\name{combine_pvals}
\alias{combine_pvals}
\title{Combine P-values}
\usage{
combine_pvals(dir)
}
\arguments{
\item{dir}{A directory that contains the results of Tukey and Dunn tests.}
}
\value{
A data frame containing the combined p-values.
}
\description{
A function to combine p-values from the functions "aov_tukey" and "kw_dunn".
All input CSV files should be located in a folder where each file has a prefix "tk_" for Tukey test results and "dn_" for Dunn test.
Each file contains the results of a test on a certain character, for example, "tk_SVL_results.csv."
}
\examples{
comb.pvals <- system.file("extdata", "tk_res", package = "morphR")

}
