% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kw_dunn.R
\name{kw_dunn}
\alias{kw_dunn}
\title{Kruskal-Wallis and Dunn Test}
\usage{
kw_dunn(data, grp = "Pop", write.dunn = TRUE, dir = "dn_female/dn_")
}
\arguments{
\item{data}{A data frame with population or species label in the first column, followed by morpholigcal data.}

\item{grp}{String. Column name for population or species.}

\item{write.dunn}{Logical(TRUE or FALSE). If TRUE, it writes the results of the Dunn test to csv files.}

\item{dir}{A directory where the Dunn test results will be stored.}
}
\value{
A list containing the results of Kruskal-Wallis and Dunn Test.
}
\description{
A function to perform Kruskal-Wallis test on all morphological data.
Also, if a significant result is detected, it automatically performs Dunn Test with a Bonferroni adjustment for p-values.
}
\examples{
data <- read.csv(system.file("extdata", "herp.data.csv", package = "MorphoTax"))
data$Sex <- NULL

kw.res <- kw_dunn(data, grp = "Pop", write.dunn = TRUE, dir = NULL)

## access results
kw.res$kruskal_summary
kw.res$dunn_combined

}
