% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levene_all.R
\name{levene_all}
\alias{levene_all}
\title{Levene's Test}
\usage{
levene_all(data, grp = "Pop")
}
\arguments{
\item{data}{A data frame with population or species label in the first column, followed by morpholigcal data.}

\item{grp}{String. Column name for population or species.}
}
\value{
A data frame summarizing the results of Levene's test.
}
\description{
A function to test homogeneity of variance in all morphological data using Levene's test.
}
\examples{
data <- read.csv(system.file("extdata", "herp.data.csv", package = "MorphoTax"))

levene.res <- levene_all(data, grp = "Pop")

}
