% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levene_sex.R
\name{levene_sex}
\alias{levene_sex}
\title{Levene's Test by Sex}
\usage{
levene_sex(data, sex = "Sex", char = "SVL", grp = "Pop")
}
\arguments{
\item{data}{A data frame with population or species label and sex information in the first two columns, followed by morpholigcal data.}

\item{sex}{String. Column name for sex information, specifying male and female.}

\item{char}{String. Column name for morphological character to test.}

\item{grp}{String. Column name for population or species.}
}
\value{
A data frame summarizing the results of Levene's test for each sex within a population or species.
}
\description{
A function to test homogeneity of variance in each sex within each population using Levene's test.
}
\examples{
data <- read.csv(system.file("extdata", "herp.data.csv", package = "morphR"))

levene.res <- levene_sex(data, sex = "Sex", char = "SVL", grp = "Pop")

}
