% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_diff.R
\name{pairwise_diff}
\alias{pairwise_diff}
\title{Pairwise Relative Difference}
\usage{
pairwise_diff(data, grp = "Pop", char = "TL")
}
\arguments{
\item{data}{A data frame with population or species label in the first column, followed by SVL and other data.}

\item{grp}{String indicating the column name of the population.}

\item{char}{String specifying the column name of the character.}
}
\value{
A data frame with all pairwise comparisons between populations,
including relative differences and which population had the higher mean.
}
\description{
A function to calculate pairwise relative differences between population means of a specific morphological character.
Calculated as:

rd = abs(pop1Mean - pop2Mean) / ((pop1Mean + pop2Mean) / 2) * 100
}
\examples{
data <- read.csv(system.file("extdata", "herp.data.csv", package = "MorphoTax"))

pw.diff <- pairwise_diff(data, grp = "Pop", char = "SVL")

head(pw.diff)

}
