% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot Principal Component Analysis}
\usage{
plot_pca(data, fixed.shape = NULL, point.color = NULL)
}
\arguments{
\item{data}{A data frame with population or species label in the first column, followed by morpholigcal data.}

\item{fixed.shape}{A vector specifying the point shape for each population or species. If NULL, random points will be used.}

\item{point.color}{A vector specifying the point color for each population or species. If NULL, random colors will be used}
}
\value{
A biplot showing the results of FDA.
}
\description{
A function to plot the first two dimensions of Principal Component Analysis (PCA).
}
\examples{
data <- read.csv(system.file("extdata", "herp.data.csv", package = "MorphoTax"))
data$Sex <- NULL

point.shape <- c("Luzon" = 8, "Mindanao" = 11, "Palawan" = 10)
point.color <- c("Luzon" = "#000000", "Mindanao" = "#FF7F0E", "Palawan" = "#D62728")

pca.plot <- plot_pca(data, point.color = point.color, 
                              fixed.shape = point.shape)
pca.plot

}
