% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pval_bubble.R
\name{plot_pval_bubble}
\alias{plot_pval_bubble}
\title{Plot P-values Bubble Chart}
\usage{
plot_pval_bubble(data, dot.val = TRUE, sort = TRUE, custom.sort = NULL)
}
\arguments{
\item{data}{A data frame with p-values of comparisons.}

\item{dot.val}{Logical. If TRUE, it adds p-values within the dot.}

\item{sort}{Logical.If TRUE, sorts the rows and columns.}

\item{custom.sort}{String specifying the column to use for custom sorting.}
}
\value{
A bubble plot of p-values
}
\description{
A function for creating a bubble plot that compares p-values, similar to 'plot_pval', but allows for custom sorting of comparisons.
}
\examples{
data <- read.csv(system.file("extdata", "cleaned.pvals.csv", package = "MorphoTax"))

pval.plot <- plot_pval_bubble(data, dot.val = TRUE, sort = TRUE, custom.sort = NULL)

}
