% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ratio_box.R
\name{plot_ratio_box}
\alias{plot_ratio_box}
\title{Ratio Boxplot}
\usage{
plot_ratio_box(data, char.as.num, char.as.den, grp, y.title = NULL)
}
\arguments{
\item{data}{A data frame with population or species label in the first column, followed by morpholigcal data.}

\item{char.as.num}{String. Column name for character representing the numerator.}

\item{char.as.den}{String. Column name for character representing the denominator.}

\item{grp}{String. Column name for population or species.}

\item{y.title}{Custom label for y-axis.}
}
\value{
A boxplot of ratio.
}
\description{
A function to create boxplot of character ratio.
}
\examples{
data <- read.csv(system.file("extdata", "herp.data.csv", package = "morphR"))

box <- plot_ratio_box(data, char.as.num = "OD", char.as.den = "SVL", grp = "Pop", 
               y.title = "Orbit Diameter/\nSnout–Vent Length")
box

}
