% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stack_comp.R
\name{plot_stack_comp}
\alias{plot_stack_comp}
\title{Plot Stacked Bar Chart by Comparison}
\usage{
plot_stack_comp(data, add.num = TRUE, flip = FALSE, sort = FALSE)
}
\arguments{
\item{data}{A data frame where the first column is "Comparisons" and the rest are p-values (as character).}

\item{add.num}{Logical. If TRUE, it adds the total counts on the bars.}

\item{flip}{Logical. If TRUE, the comparison label shifts to the x-axis.}

\item{sort}{Logical. If TRUE, sorts the comparison based on significane.}
}
\value{
A stacked bar plot showing the total count of significant and non-significant characters.
}
\description{
This function counts how many p-values for each comparison across are below or above the p-value threshold of 0.05 and
then plots a stacked bar chart to visualize the result.
}
\examples{
data <- read.csv(system.file("extdata", "cleaned.pvals.csv", package = "MorphoTax"))

pval.stack <- plot_stack_comp(data, add.num = TRUE, flip = FALSE, sort = TRUE)

pval.stack

}
