% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stack_morph.R
\name{plot_stack_morph}
\alias{plot_stack_morph}
\title{Plot Stacked Bar Chart by Characters}
\usage{
plot_stack_morph(data, add.num = TRUE, flip = FALSE, sort = FALSE)
}
\arguments{
\item{data}{A data frame where the first column is "Comparisons" and the rest are p-values (as character).}

\item{add.num}{Logical. If TRUE, it adds the total counts on the bars.}

\item{flip}{Logical. If TRUE, the character labels shifts to the x-axis.}

\item{sort}{Logical. If TRUE, sorts the character labels based on significance.}
}
\value{
A stacked bar plot showing the total count of significant and non-significant characters.
}
\description{
This function counts how many p-values for each character across comparisons are below or above the p-value threshold of 0.05
and then plots a stacked bar chart to visualize the results.
}
\examples{
data <- read.csv(system.file("extdata", "cleaned.pvals.csv", package = "MorphoTax"))

pval.stack <- plot_stack_morph(data, add.num = TRUE, flip = FALSE, sort = TRUE)

pval.stack

}
