% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_juv.R
\name{remove_juv}
\alias{remove_juv}
\title{Remove Juveniles}
\usage{
remove_juv(data, char = "SVL", threshold = 70, sum = TRUE)
}
\arguments{
\item{data}{A data frame with population or species label in the first column, followed by morpholigcal data.}

\item{char}{String. Column name of the character to use, typically snout–vent length (SVL).}

\item{threshold}{Numeric. A value indicating the character value to be considered as adult.}

\item{sum}{Logical. If TRUE, it prints the number and name of samples removed.}
}
\value{
A filtered data frame.
}
\description{
A function to remove juveniles based on specified threshold.
}
\examples{
data <- read.csv(system.file("extdata", "herp.data.csv", package = "MorphoTax"))

filtered.data <- remove_juv(data, char = "SVL", threshold = 70, sum = TRUE)

filtered.data

}
