% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_non_bin.R
\name{remove_non_bin}
\alias{remove_non_bin}
\title{Remove Samples with Non-binary Values}
\usage{
remove_non_bin(data, grp = NULL)
}
\arguments{
\item{data}{A data frame to filter.}

\item{grp}{String. Name of group or population to report number of removed samples.}
}
\value{
A filtered data frame with binary data.
}
\description{
A function to remove samples with non-binary values.
}
\examples{
data <- read.csv(system.file("extdata", "herp.data.csv", package = "MorphoTax"))

filtered.data <- remove_non_bin(data, grp = "Pop")

head(filtered.data)

}
