% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_non_num.R
\name{remove_non_num}
\alias{remove_non_num}
\title{Remove Non-Numeric Samples}
\usage{
remove_non_num(data, cols.range = 3:ncol(data), sum = TRUE)
}
\arguments{
\item{data}{A data frame with population or species label in the first column, followed by morpholigcal data.}

\item{cols.range}{Vector. Column range to check for non-numeric or empty cells. Input can be "3:6" or "3:ncol(data)". Default 3:ncol(data).}

\item{sum}{Logical. If TRUE, it prints the number and name of samples removed.}
}
\value{
A filtered data frame.
}
\description{
A function to remove samples (rows) with non-numeric or empty cells.
}
\examples{
data <- read.csv(system.file("extdata", "herp.data.csv", package = "MorphoTax"))

filtered.data <- remove_non_num(data, cols.range = 3:ncol(data), sum = TRUE)

filtered.data

}
