% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove Outliers}
\usage{
remove_outliers(data, char = "SVL", grp = NULL, q1 = 0.25, q3 = 0.75)
}
\arguments{
\item{data}{A data frame with population or species label in the first column, followed by morpholigcal data.}

\item{char}{String. Column name for character to use for removing outliers.}

\item{grp}{String. Column name for population or species. If NULL, the entire dataset will be considered as one population.}

\item{q1}{Numeric. Value for lower quartile. Default is set to 0.25.}

\item{q3}{Numeric. Value for upper quartile. Default is set to 0.75.}
}
\value{
A filtered data frame.
}
\description{
A function to remove outliers based on interquartile range method.
It uses only one character as a reference in removing samples.
}
\examples{
data <- read.csv(system.file("extdata", "herp.data.csv", package = "MorphoTax"))

filtered.data <- remove_outliers(data, char = "SVL", grp = "Pop", q1 = 0.25, q3 = 0.75)

filtered.data

}
