% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_pop.R
\name{remove_pop}
\alias{remove_pop}
\title{Remove Populations}
\usage{
remove_pop(data, grp = "Pop", threshold = 4)
}
\arguments{
\item{data}{A data frame with population or species label in the first column, followed by morpholigcal data.}

\item{grp}{String. Column name for population or species.}

\item{threshold}{Numeric. Number of samples required for each population to be retained.
Populations with fewer than this number will be removed. Default is 4.}
}
\value{
A filtered data frame.
}
\description{
A function to remove populations with small samples or individuals.
}
\examples{
data <- read.csv(system.file("extdata", "herp.data.csv", package = "MorphoTax"))

filtered.data <- remove_pop(data, grp = "Pop", threshold = 4)

filtered.data

}
