% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fda.R
\name{run_fda}
\alias{run_fda}
\title{Flexible Discriminant Analysis}
\usage{
run_fda(data, con.mat = TRUE, accu = TRUE, coef = TRUE, split = NULL)
}
\arguments{
\item{data}{A data frame with population or species label in the first column, followed by morpholigcal data.}

\item{con.mat}{Logical (TRUE or FALSE). If TRUE, it writes the confusion matrix to a csv file.}

\item{accu}{Logical (TRUE or FALSE). If TRUE, it calculates and prints the classification accuracy.}

\item{coef}{Logical (TRUE or FALSE). If TRUE, it writes the coefficient values to a csv file.}

\item{split}{Numeric. Proportion of data to be assigned as training data. If NULL, all data will be used as training set.}
}
\value{
A list containing the FDA results.
}
\description{
A function to perform Flexible Discriminant Analysis (FDA), a classification method that integrates non-linear regression with
linear discriminant analysis. The function uses the 'fda' function from the 'mda' package. It then writes the confusion matrix,
extracts the coefficients, and calculates the accuracy based on confusion matrix.
}
\examples{
data <- read.csv(system.file("extdata", "herp.data.csv", package = "morphR"))
data$Sex <- NULL

fda.res <- run_fda(data, con.mat = TRUE, coef = TRUE, accu = TRUE, split = 0.7)

fda.res

}
