% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pca.R
\name{run_pca}
\alias{run_pca}
\title{Principal Component Analysis}
\usage{
run_pca(data, prop.var = TRUE, var.load = TRUE, pts.load = TRUE)
}
\arguments{
\item{data}{A data frame with population or species label in the first column, followed by morpholigcal data.}

\item{prop.var}{Logical (TRUE or FALSE). If TRUE, it calculates the varaince explained by the first two components.}

\item{var.load}{Logical (TRUE or FALSE). If TRUE, it extracts and writes the varaible loadings to csv files.}

\item{pts.load}{Logical (TRUE or FALSE). If TRUE, it extracts and writes the point loadings to csv files .}
}
\value{
A list containing the FDA results.
}
\description{
A function to perform Principal Component Analysis (PCA), a  method used to reduce the dimensionality of data by determining
the principal components (directions) that maximizes the variance in the dataset. The function uses the 'pincomp' function. It then caculates
the variance explained of the first two components, extracts the variable and point loadings and writes them to csv files.
}
\examples{
data <- read.csv(system.file("extdata", "herp.data.csv", package = "morphR"))
data$Sex <- NULL

pca.res <- run_pca(data, prop.var = TRUE, var.load = TRUE, pts.load = TRUE)

pca.res

}
