% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pca_svm.R
\name{run_pca_svm}
\alias{run_pca_svm}
\title{Support Vector Machine on PCA}
\usage{
run_pca_svm(data, point.shape = NULL, class.color = NULL, pop.order = NULL)
}
\arguments{
\item{data}{A data frame with population or species label in the first column, followed by morpholigcal data.}

\item{point.shape}{A vector specifying the point shapes for populations. If NULL, it assigns random point shapes.}

\item{class.color}{A vector specifying the point and class colors. If NULL, it assigns random point shapes and colors.}

\item{pop.order}{A vector specifying the order of populations to be shown in the plot legend.}
}
\value{
A plot showing the svm boundary for predicted classes.
}
\description{
A function to perform support vector machine (SVM) on the first two dimensions of Principal Component Analysis.
SVM is a classification algorithm that partitions samples into classes by searching the optimal hyperplane margin
that maximizes the boundary among classes (in this case, putative population or species) in two or more dimensions.
The function uses the 'svm' function from the package 'e1071', implementing the 'svmLinear'method and a 10-fold cross-validation.
The function ouputs a plot showing the svm boundary for predicted classes.
}
\examples{
data <- read.csv(system.file("extdata", "herp.data.csv", package = "morphR"))
data$Sex <- NULL

point.shape <- c("Luzon" = 8, "Mindanao" = 11, "Palawan" = 10)
point.color <- c("Luzon" = "#000000", "Mindanao" = "#FF7F0E", "Palawan" = "#D62728")
pop <- c("Palawan", "Mindanao", "Luzon")

pca.svm <- run_pca_svm(data, point.shape = point.shape, class.color = NULL, pop.order = pop)

pca.svm

}
