% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapiro_all.R
\name{shapiro_all}
\alias{shapiro_all}
\title{Shapiro-Wilk Test}
\usage{
shapiro_all(data, grp = "Pop")
}
\arguments{
\item{data}{A data frame with population or species label in the first column, followed by morpholigcal data.}

\item{grp}{String. Column name for population or species.}
}
\value{
A data frame summarizing the results of Shapiro-Wilk Test.
}
\description{
A function to test normality of all morphological data using Shapiro-Wilk Test.
}
\examples{
data <- read.csv(system.file("extdata", "herp.data.csv", package = "MorphoTax"))
data$Sex <- NULL

shapiro.res <- shapiro_all(data, grp = "Pop")

shapiro.res

}
