% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapiro_sex.R
\name{shapiro_sex}
\alias{shapiro_sex}
\title{Shapiro-Wilk Test by Sex}
\usage{
shapiro_sex(data, sex = "Sex", grp = "Pop")
}
\arguments{
\item{data}{A data frame with population or species label and sex information in the first two columns, followed by morpholigcal data.}

\item{sex}{String. Column name for sex information, specifying male and female.}

\item{grp}{String. Column name for population or species.}
}
\value{
A data frame summarizing the results of Shapiro-Wilk Test for each sex within a population or species.
}
\description{
A function to test normality of data in each sex within each population using Shapiro-Wilk Test.
}
\examples{
data <- read.csv(system.file("extdata", "herp.data.csv", package = "morphR"))

shapiro.res <- shapiro_sex(data, sex = "Sex", grp = "Pop")

shapiro.res

}
