% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_morph.R
\name{summarize_morph}
\alias{summarize_morph}
\title{Summarize Morphological Data}
\usage{
summarize_morph(data, grp = NULL)
}
\arguments{
\item{data}{A data frame with population or species label and sex information in the first two columns, followed by morpholigcal data.}

\item{grp}{String. Column name for population or species.}
}
\value{
A data frame containing summary statistics of all morphological data for every population or species.
}
\description{
A function to get the summary statistics of all morphological data.
}
\examples{
data <- read.csv(system.file("extdata", "herp.data.csv", package = "MorphoTax"))
data$Sex <- NULL

sum.res <- esult <- summarize_morph(data, grp = "Pop")

head(sum.res)

}
