% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_ratio.R
\name{summarize_ratio}
\alias{summarize_ratio}
\title{Summarize Character Ratio}
\usage{
summarize_ratio(data, char.as.num, char.as.den, grp)
}
\arguments{
\item{data}{A data frame with population or species label and sex information in the first two columns, followed by morpholigcal data.}

\item{char.as.num}{String. Column name for character representing the numerator.}

\item{char.as.den}{String. Column name for character representing the denominator.}

\item{grp}{String. Column name for population or species.}
}
\value{
A data frame containing summary of ratio.
}
\description{
A function to calculate and summarize the ratio of two morphological data. The summary includes the number of samples with longer length
(ratio >1.0) and shorter length (ratio <1.0), percentages of samples relative to the total, range (min–max), mean, and the calculated
relative difference (e.g., |ED - SVL|/SV) * 100).
}
\examples{
data <- read.csv(system.file("extdata", "herp.data.csv", package = "morphR"))

ratio.sum <- summarize_ratio(data, char.as.num = "TD", char.as.den = "SVL", grp = "Pop")

head(ratio.sum)

}
