% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_sex.R
\name{t_test_sex}
\alias{t_test_sex}
\title{T-test by Sex}
\usage{
t_test_sex(data, sex = "Sex", char = "SVL", grp = "Pop")
}
\arguments{
\item{data}{A data frame with population or species label and sex information in the first two columns, followed by morpholigcal data.}

\item{sex}{String. Column name for sex information, specifying male and female.}

\item{char}{String. Column name for morphological character to test.}

\item{grp}{String. Column name for population or species.}
}
\value{
A data frame containing the results of T-test.
}
\description{
A function to test differences in characters between sexes with each population using independent T-test.
}
\examples{
data <- read.csv(system.file("extdata", "herp.data.csv", package = "morphR"))

t.test.res <- t_test_sex(data, sex = "Sex", char = "SVL", grp = "Pop")

head(t.test.res)

}
