% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_test_sex.R
\name{wilcox_test_sex}
\alias{wilcox_test_sex}
\title{Wilcoxon Test by Sex}
\usage{
wilcox_test_sex(data, sex = "Sex", char = "SVL", grp = "Pop", all = FALSE)
}
\arguments{
\item{data}{A data frame with population or species label and sex information in the first two columns, followed by morpholigcal data.}

\item{sex}{String. Column name for sex information, specifying male and female.}

\item{char}{String. Column name for morphological character to test.}

\item{grp}{String. Column name for population or species.}

\item{all}{Logical. If TRUE, the Wilcoxon Test will be performed to all character or numeric columns.}
}
\value{
A data frame containing the results of Wilcoxon test.
}
\description{
A function to test differences in characters between sexes within each population using Wilcoxon Test.
}
\examples{
data <- read.csv(system.file("extdata", "herp.data.csv", package = "MorphoTax"))

wilcox.res <- wilcox_test_sex(data, sex = "Sex", char = "SVL", grp = "Pop", all = FALSE)

head(wilcox.res)

}
