% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_errorbar.R
\name{plot_errorbar}
\alias{plot_errorbar}
\title{Plot Error Bar}
\usage{
plot_errorbar(
  data,
  char = "SVL",
  sex = "Sex",
  grp = "Pop",
  test = c("t-test", "wilcox", "none"),
  pop.order = NULL,
  asterisk = TRUE
)
}
\arguments{
\item{data}{A data frame with population or species label and sex information in the first two columns, followed by morpholigcal data.}

\item{char}{String. Column name for morphological character to test.}

\item{sex}{String. Column name for sex information, specifying male and female.}

\item{grp}{String. Column name for population or species.}

\item{test}{String. Specify which test to use ("t-test", "wilcox", "none").}

\item{pop.order}{A vector specifying the order of populations to be shown in the plot legend.}

\item{asterisk}{Logical. If TRUE, it adds an asterisk to the plot to indicate a significant result.}
}
\description{
A function for creating an error bar plot and test differences in characters between
sexes within each population using the T-test and Wilcoxon test.
}
\examples{
data <- read.csv(system.file("extdata", "herp.data.csv", package = "MorphoTax"))
pop <- c("Luzon", "Palawan", "Mindanao")

svl.bar <- plot_errorbar(data, char = "SVL", sex = "Sex", 
             grp = "Pop", test = "t-test", pop.order = pop, asterisk = TRUE)

svl.bar

}
