% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pval.R
\name{plot_pval}
\alias{plot_pval}
\title{Plot P-values}
\usage{
plot_pval(data, dot.val = TRUE)
}
\arguments{
\item{data}{A data frame with p-values of comparisons.}

\item{dot.val}{Logical. If TRUE, it adds p-values within the dot.}
}
\description{
A function for creating a dot plot to compare p-values.
}
\examples{
data <- read.csv(system.file("extdata", "cleaned.pvals.csv", package = "MorphoTax"))

pval_plot <- plot_pval(data, dot.val = TRUE)

}
